+++ common/src/main/java/io/github/flemmli97/runecraftory/common/lib/LibEntities.java
@@ -69,6 +69,9 @@ public class LibEntities {
     public static final ResourceLocation darkBall = new ResourceLocation(RuneCraftory.MODID, "dark_ball");
     public static final ResourceLocation darkness = new ResourceLocation(RuneCraftory.MODID, "darkness");
 
+    public static final ResourceLocation runey = new ResourceLocation(RuneCraftory.MODID, "runey");
+    public static final ResourceLocation runeOrb = new ResourceLocation(RuneCraftory.MODID, "rune_orb");
+
     public static final ResourceLocation butterflySummoner = new ResourceLocation(RuneCraftory.MODID, "butterfly_summoner");
 
     public static final ResourceLocation fishingHook = new ResourceLocation(RuneCraftory.MODID, "fishing_hook");