+++ fabric/src/main/java/io/github/flemmli97/runecraftory/fabric/config/ConfigHolder.java
@@ -42,6 +42,7 @@ public record ConfigHolder<T>(T configSpec, Consumer<T> loader) {
         GeneralConfig.disableItemStatSystem = spec.disableItemStatSystem.get();
         GeneralConfig.disableCropSystem = spec.disableCropSystem.get();
         GeneralConfig.witherChance = spec.witherChance.get().floatValue();
+        GeneralConfig.runeyChance = spec.runeyChance.get().floatValue();
         GeneralConfig.seasonedSnow = spec.seasonedSnow.get();
         GeneralConfig.allowLockedCrafting = spec.allowLockedCrafting.get();