+++ forge/src/main/java/io/github/flemmli97/runecraftory/forge/config/GeneralConfigSpec.java
@@ -32,6 +32,7 @@ public class GeneralConfigSpec {
     public final ForgeConfigSpec.BooleanValue disableItemStatSystem;
     public final ForgeConfigSpec.BooleanValue disableCropSystem;
     public final ForgeConfigSpec.DoubleValue witherChance;
+    public final ForgeConfigSpec.DoubleValue runeyChance;
     public final ForgeConfigSpec.BooleanValue seasonedSnow;
     public final ForgeConfigSpec.BooleanValue allowLockedCrafting;
 
@@ -94,6 +95,7 @@ public class GeneralConfigSpec {
         this.disableItemStatSystem = builder.comment("If true item get no stats assigned as defined per datapack. ", "Note: Weapons will then do no damage").define("Disable Item Stat System", GeneralConfig.disableItemStatSystem);
         this.disableCropSystem = builder.comment("If true crop data will be disabled.", "Note: Crops from this mod will not function anymore").define("Disable Crop System", GeneralConfig.disableCropSystem);
         this.witherChance = builder.comment("Change for a crop to wither if its not been watered", "If crop is already withered it and it doesnt get watered it will turn into withered grass").defineInRange("Wither Chance", GeneralConfig.witherChance, 0, 1);
+        this.runeyChance = builder.comment("Chance for a runey to spawn when harvesting fully grown crops").defineInRange("Runey Chance", GeneralConfig.runeyChance, 0, 1);
         this.seasonedSnow = builder.comment("If biome temperature should be adjusted based on current season. Can cause snowfall during winter").define("Seasoned Biome Temp", GeneralConfig.seasonedSnow);
         this.allowLockedCrafting = builder.comment("If true allows crafting of not unlocked recipes. They will cost more rp to craft").define("Allow Locked Crafting", GeneralConfig.allowLockedCrafting);
         builder.pop();