+++ Changelog.md
@@ -12,6 +12,10 @@ RuneCraftory 1.4.0
   the item directly but without an duplicate item penalty.  
   Gui for crafting got redesigned to make choosing recipes easier.  
   Crafting an item without a valid recipe now yields something different
+- Fix biome tags
+- Disable spell keys when gui is open
+- Runeys and rune orbs can now spawn when harvesting fully grown crops. They will refresh your runepoints and level a random skill
+  or a random stat
 
 RuneCraftory 1.3.4
 ================